Completed File Backup Program
Completed File Backup program that takes whatever files are in the current directory and subdirectories and zips them into a folder with the name of the directory, and the version number.

I deviated from the guide a little more than usual for this project, because I saw this program as particularly more useful than the past exercises. I wanted to make the program a little more versatile; working in whatever directory, no matter the path.

My next step, before moving on to the next exercise will be to turn this python program into an executable using cx_Freeze.